% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comment_block}
\alias{comment_block}
\title{Format text as a code block comment}
\usage{
comment_block(text, width = 80L, char = "#", pad = 1L)
}
\arguments{
\item{text}{Character string to format.}

\item{width}{Total width of the comment block. Default 80.}

\item{char}{Character to use for the border. Default "#".}

\item{pad}{Number of spaces to pad the text. Default 1.}
}
\value{
Character string formatted as a comment block.
}
\description{
Formats text as a monospace-style code comment block. Useful for
creating visually distinct section headers in scripts.
}
\examples{
# Create a section header
cat(comment_block("Data Import"))
# ############################################################################
# #                              Data Import                                 #
# ############################################################################

# Shorter width
cat(comment_block("Setup", width = 40))

}
