% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comment_header}
\alias{comment_header}
\title{Format text as a simple comment header}
\usage{
comment_header(text, char = "-", width = 80L)
}
\arguments{
\item{text}{Character string to format.}

\item{char}{Character to use for decoration. Default "-".}

\item{width}{Total width of the header. Default 80.}
}
\value{
Character string formatted as a comment header.
}
\description{
Formats text as a simple single-line comment header.
}
\examples{
cat(comment_header("Load Libraries"))
# # Load Libraries --------------------------------------------------------

}
