% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_test}
\alias{db_test}
\title{Test database connection}
\usage{
db_test(conn)
}
\arguments{
\item{conn}{A DBI connection object.}
}
\value{
TRUE if connection is valid, FALSE otherwise.
}
\description{
Verifies that a database connection is valid.
}
\examples{
\dontrun{
conn <- db_connect(driver = "sqlite", database = "test.db")
db_test(conn)
#> TRUE
}

}
