% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{fundr_palette}
\alias{fundr_palette}
\title{fundr palettes}
\usage{
fundr_palette(palette = c("primary", "secondary", "tertiary"))
}
\arguments{
\item{palette}{Palette name: "primary", "secondary", or "tertiary".}
}
\value{
An unnamed character vector of hex color codes.
}
\description{
Returns a vector of hex colors from one of the fundr palettes.
}
\examples{
# Get the primary palette (2 colors)
fundr_palette("primary")

# Get the secondary palette (10 colors)
fundr_palette("secondary")

# Get the tertiary palette (4 colors)
fundr_palette("tertiary")

}
\seealso{
Other colors: 
\code{\link{fundr_colors}()},
\code{\link{fundr_pal}()},
\code{\link{scale_colour_fundr}()},
\code{\link{scale_fill_fundr}()}
}
\concept{colors}
