% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal_year.R
\name{fy_quarter}
\alias{fy_quarter}
\title{Fiscal quarter for a date}
\usage{
fy_quarter(date, fy_start_month = getOption("fundr.fy_start_month", 7L))
}
\arguments{
\item{date}{A Date (or something coercible via as.Date()).}

\item{fy_start_month}{Integer 1-12. Default uses \code{getOption("fundr.fy_start_month", 7)}.}
}
\value{
Integer vector (1-4).
}
\description{
Quarters are numbered 1-4 within the fiscal year defined by \code{fy_start_month}.
}
\examples{
# July fiscal year: Q1 starts in July
fy_quarter(as.Date("2024-07-15"))
fy_quarter(as.Date("2024-10-15"))

# Vectorized
dates <- as.Date(c("2024-07-01", "2024-10-01", "2025-01-01", "2025-04-01"))
fy_quarter(dates)

}
\seealso{
Other fiscal-year: 
\code{\link{fy_label}()},
\code{\link{fy_year}()}
}
\concept{fiscal-year}
