% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{notin}
\alias{notin}
\alias{\%notin\%}
\alias{\%!in\%}
\title{Not in operator}
\usage{
x \%notin\% table

x \%!in\% table
}
\arguments{
\item{x}{Vector of values to check.}

\item{table}{Vector of values to check against.}
}
\value{
Logical vector the same length as \code{x}.
}
\description{
The negation of the \code{\%in\%} operator. Returns TRUE for elements of \code{x}
that are not in \code{table}.
}
\examples{
# Find values not in a set
c(1, 2, 3, 4, 5) \%notin\% c(2, 4)
#> TRUE, FALSE, TRUE, FALSE, TRUE

# Filter data frame rows
df <- data.frame(status = c("Active", "Lapsed", "Active", "Never"))
df[df$status \%notin\% c("Lapsed", "Never"), ]

}
