% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{set_fy_start_month}
\alias{set_fy_start_month}
\title{Set default fiscal year start month}
\usage{
set_fy_start_month(month)
}
\arguments{
\item{month}{Integer 1-12 indicating fiscal year start month.}
}
\value{
Invisibly returns the previous value.
}
\description{
Convenience function to change the default fiscal year start month
without calling \code{fundr_setup()}.
}
\examples{
# Save current setting
old_month <- getOption("fundr.fy_start_month")

# Set to October (federal fiscal year)
set_fy_start_month(10)
get_fy_start_month()

# Restore previous setting
if (!is.null(old_month)) options(fundr.fy_start_month = old_month)

}
