% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_convert.R
\name{to_camel_case}
\alias{to_camel_case}
\title{Convert strings to camelCase}
\usage{
to_camel_case(x)
}
\arguments{
\item{x}{Character vector to convert.}
}
\value{
Character vector in camelCase.
}
\description{
Converts strings from snake_case or other formats to camelCase.
}
\examples{
to_camel_case("first_name")
#> "firstName"

to_camel_case("gift_amount")
#> "giftAmount"

}
