% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_convert.R
\name{to_snake_case}
\alias{to_snake_case}
\title{Convert strings to snake_case}
\usage{
to_snake_case(x)
}
\arguments{
\item{x}{Character vector to convert.}
}
\value{
Character vector in snake_case.
}
\description{
Converts strings from various formats (TitleCase, camelCase, etc.)
to snake_case. Useful for standardizing column names.
}
\examples{
to_snake_case("FirstName")
#> "first_name"

to_snake_case("firstName")
#> "first_name"

to_snake_case(c("GiftAmount", "DonorID", "LastGiftDate"))
#> "gift_amount", "donor_id", "last_gift_date"

# Already snake_case passes through
to_snake_case("gift_amount")
#> "gift_amount"

}
