% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundr-package.R
\docType{package}
\name{fundr-package}
\alias{fundr}
\alias{fundr-package}
\title{fundr: Fundraising analytics helpers}
\description{
Utility functions and reference tools for fundraising and advancement analytics,
designed to be tidyverse-compatible with minimal required dependencies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mattfarrow1/fundr}
  \item \url{https://mattfarrow1.github.io/fundr/}
  \item Report bugs at \url{https://github.com/mattfarrow1/fundr/issues}
}

}
\author{
\strong{Maintainer}: Matt Farrow \email{matt@mattfarrow.co} [copyright holder]

Other contributors:
\itemize{
  \item Claude (Anthropic AI) (AI programming assistant) [contributor]
}

}
\keyword{internal}
