% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{fundr_pal}
\alias{fundr_pal}
\title{Palette function for ggplot2 discrete scales}
\usage{
fundr_pal(palette = c("primary", "secondary", "tertiary"), direction = 1)
}
\arguments{
\item{palette}{Palette name: "primary", "secondary", or "tertiary".}

\item{direction}{If 1, use palette order; if -1, reverse.}
}
\value{
A function that takes \code{n} and returns \code{n} colors.
}
\description{
Palette function for ggplot2 discrete scales
}
