% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{fundr_palette}
\alias{fundr_palette}
\title{fundr palettes}
\usage{
fundr_palette(palette = c("primary", "secondary", "tertiary"))
}
\arguments{
\item{palette}{Palette name: "primary", "secondary", or "tertiary".}
}
\value{
An unnamed character vector of hex color codes.
}
\description{
fundr palettes
}
