% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal_year.R
\name{fy_quarter}
\alias{fy_quarter}
\title{Fiscal quarter for a date}
\usage{
fy_quarter(date, fy_start_month = getOption("fundr.fy_start_month", 7L))
}
\arguments{
\item{date}{A Date (or something coercible via as.Date()).}

\item{fy_start_month}{Integer 1-12. Default uses \code{getOption("fundr.fy_start_month", 7)}.}
}
\value{
Integer vector (1-4).
}
\description{
Quarters are numbered 1-4 within the fiscal year defined by \code{fy_start_month}.
}
