% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get_fy_start_month}
\alias{get_fy_start_month}
\title{Get default fiscal year start month}
\usage{
get_fy_start_month()
}
\value{
Integer 1-12 representing the fiscal year start month.
}
\description{
Returns the fiscal year start month from options, or the package default
(July = 7) if not set.
}
\examples{
# Get current default
get_fy_start_month()

# After setting a different default
fundr_setup(fy_start_month = 1, quiet = TRUE)
get_fy_start_month()
#> 1

}
