% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_fundr}
\alias{scale_fill_fundr}
\title{Discrete fill scale using fundr palettes}
\usage{
scale_fill_fundr(
  palette = c("primary", "secondary", "tertiary"),
  direction = 1,
  ...
)
}
\arguments{
\item{palette}{Palette name: "primary", "secondary", or "tertiary".}

\item{direction}{If 1, use palette order; if -1, reverse.}

\item{...}{Passed to ggplot2::discrete_scale().}
}
\description{
Discrete fill scale using fundr palettes
}
